geo: fix three faults the first RDAP fill exposed
1. MAINTAINER OBJECTS RECORDED AS REGISTRANTS

Several entities can carry the RDAP registrant role and the code took the first in
document order. Verified against live RIPE for 51.180.0.0 - 51.181.255.255, three
do: MNT-ADSI (a maintainer), ORG-ARI3-RIPE with fn "A100 ROW Inc" (the actual
holder), and RIPE-NCC-HM-MNT (RIPE's hostmaster maintainer). Document order put
the maintainer first, so 26 allocations were recorded as MNT-ADSI. APPLE-MNT,
MICROSOFT-MAINT, ORCL-MNT and OPENDNS-MNT came from the same fault.

A maintainer is an access-control object, not a holder, and its name matches none
of the ARIN-style names any exclusion pattern is written against. The
discriminator that generalises is that a maintainer's formatted name equals its
handle, while a real organisation has a human name distinct from it. Handle shape
is a second signal. Verified: MNT-ADSI now yields "A100 ROW Inc" and APPLE-MNT
yields "Apple Inc.".

2. WHOLE /8 PLACEHOLDERS RECORDED AS ALLOCATIONS

ARIN answers 303 for early-registration space and redirects to the responsible
registry. For 151.220.0.1 that lands on RIPE, which returns the entire
151.0.0.0/8 as RIPE-NCC-MANAGED-ADDRESS-BLOCK with no entities at all. The code
fell back to the network name and stored a 16,777,216-address allocation whose
registrant is administrative filler. Worse than storing nothing: the coverage test
then treats every candidate inside as answered and never queries it again. Two
such rows, for 151.0.0.0/8 and 46.0.0.0/8, suppressed roughly 33.5 million
addresses of candidates.

Placeholder names are now recognised and stored with a NULL registrant and a note
reading placeholder:NAME. apply_splits treats a cache row with a NULL registrant
as KNOWN rather than unknown, because the registry was asked and holds no
assignment, which is positive evidence that no DoD registration exists. Treating
it as unknown would defer such ranges permanently, since no later query can
produce a better answer. Verified end to end: DoD excluded, no-cache-row deferred,
real registrant decomposed, placeholder-covered decomposed.

3. round(double precision, integer) DOES NOT EXIST

Section 6 of decomposition_exclusions.sql aborted with that error, because 2
raised to an integer power returns double precision in PostgreSQL. The base is now
numeric. The other SQL in the repo was already safe: it casts pow() to bigint
inside the sum. Also reorders the registrant report so matched rows appear first;
the previous ordering with LIMIT 60 hid every DoD match on a candidate set with
more than 60 distinct registrants.

Adds rdap_registrant_repair.sql to delete the rows written by faults 1 and 2 so
they are re-fetched correctly. Safe: the table is a cache, nothing else writes to
it, and no probe, geolocation or targeting row is touched.

+++ geo/decomposition_exclusions.sql
@@ -238,10 +238,15 @@ LIMIT 50;
 -- row here except the DoD ones is work that will actually be done.
 -- ---------------------------------------------------------------------------
 
+-- 2 raised to an integer power returns double precision in PostgreSQL, and there
+-- is no round(double precision, integer). The base is therefore numeric, which
+-- keeps the sum numeric and lets round take a scale argument. Getting this wrong
+-- aborts the whole statement with "function round(double precision, integer) does
+-- not exist", which is how it was found.
 SELECT cov.registrant,
        count(*)                                                     AS wide_ranges,
-       sum(2 ^ (24 - masklen(d.network)))::bigint                     AS sub24_rows,
-       round(sum(2 ^ (24 - masklen(d.network))) / 64800.0, 2)          AS probe_days,
+       sum(2::numeric ^ (24 - masklen(d.network)))::bigint             AS sub24_rows,
+       round(sum(2::numeric ^ (24 - masklen(d.network))) / 64800.0, 2) AS probe_days,
        min(masklen(d.network))                                        AS widest,
        EXISTS (SELECT 1 FROM decomposition_exclusions x
                 WHERE x.active AND cov.registrant ILIKE x.operator_pattern)
